refactor(engine-core): passing the renderer from an import statement in compiled templates
